tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-state-snapped-layout-change.html

<!DOCTYPE html>
<title>@container: scroll-state(snapped) layout change</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#snapped">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  :root {
    scroll-snap-type: block proximity;
  }
  body {
    margin: 0;
  }
  #filler-before {
    height: 10px;
  }
  #filler-after {
    height: 10000px;
  }
  #snapped {
    position: relative;
    top: 3000px; /* Should be enough to not snap for proximity */
    container-type: scroll-state;
    scroll-snap-align: start;
    width: 100px;
    height: 100px;
  }
  #target {
    --snapped: no;
    @container scroll-state(snapped) {
      --snapped: yes;
    }
  }
</style>
<div id="filler-before"></div>
<div id="snapped" class="snapped">
  <span id="target">#target</span>
</div>
<div id="filler-after"></div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "no");
  }, "Check scroll-state(snapped) initially not matching");

  promise_test(async t => {
    snapped.style.top = "auto";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "yes");
  }, "scroll-state(snapped) matching after relative position changed");

</script>
